Issue #11163: Enforced FileSize LeaveToken
