+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/abstractjavadoc/InputAbstractJavadocLeaveTokenOne.java
@@ -1,0 +1,108 @@
+/*
+com.puppycrawl.tools.checkstyle.checks.javadoc.AbstractJavadocCheckTest$JavadocVisitLeaveCheck
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.abstractjavadoc;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/////////////
+// CLASSES //
+/////////////
+
+/**Javadoc*/
+class/**nope*/ InputAbstractJavadocLeaveTokenOne/**nope*/{
+    /**Javadoc*/
+    protected/**nope*/ class/**nope*/ B/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    private/**nope*/ static/**nope*/ class/**nope*/ C/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    @Component/**nope*/ class/**nope*/ D/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    @Component/**nope*/ private/**nope*/ class/**nope*/ E/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    private/**nope*/ @Component/**nope*/ class/**nope*/ F/**nope*/{/**nope*/}
+}
+
+//////////////////
+// CONSTRUCTORS //
+//////////////////
+
+/**Javadoc*/
+class/**nope*/ AA1/**nope*/{
+    /**Javadoc*/
+    AA1/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+/**Javadoc*/
+class/**nope*/ BB1/**nope*/{
+    /**Javadoc*/
+    private/**nope1*/ BB1/**nope2*/(/**nope3*/)/**nope4*/{/**nope5*/}/**nope6*/
+}
+
+class/**nope*/ DD1/**nope*/{
+    /**Javadoc*/
+    @Component/**nope*/ DD1/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class/**nope*/ EE1/**nope*/{
+    /**Javadoc*/
+    @Component/**nope*/ private/**nope*/ EE1/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class/**nope*/ FF1/**nope*/{
+    /**Javadoc*/
+    private/**nope*/ @Component/**nope*/ FF1/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+/////////////
+// METHODS //
+/////////////
+
+class AAA1 {
+    /**Javadoc*/
+    void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class BBB1 {
+    /**Javadoc*/
+    private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class CCC1 {
+    /**Javadoc*/
+    static/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class DDD1 {
+    /**Javadoc*/
+    @Component/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class EEE1 {
+    /**Javadoc*/
+    @Component/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class FFF1 {
+    /**Javadoc*/
+    static/**nope*/ @Component/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class GGG1 {
+    /**Javadoc*/
+    void/**nope*/ a/**nope*/(@Component/**nope*/int/**nope*/ a/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class HHH1 {
+    /**Javadoc*/
+    java.lang.String/**nope*/ a/**nope*/()/**nope*/{/**nope*/return null;/**nope*/}/**nope*/
+}
